Revert "change pass"
