--- SEP490_Ezstay/AccountAPI/DTO/Response/ChangePasswordResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace AccountAPI.DTO.Response
-{
-    public class ChangePasswordResponse
-    {
-        public bool Success { get; set; }
-        public string Message { get; set; } = string.Empty;
-    }
-}
