--- SEP490_Ezstay/AccountAPI/DTO/Resquest/ChangePasswordRequest.cs
@@ -1,9 +1,0 @@
-﻿namespace AccountAPI.DTO.Resquest
-{
-    public class ChangePasswordRequest
-    {
-       public string? Email { get; set; } = null!;
-        public string OldPassword { get; set; } = null!;
-        public string NewPassword { get; set; } = null!;
-    }
-}
