+++ SEP490_Ezstay/AccountAPI/Service/AuthApiClient.cs
@@ -1,5 +1,4 @@
 ﻿using AccountAPI.DTO.Response;
-using AccountAPI.DTO.Resquest;
 using AccountAPI.Service.Interfaces;
 
 namespace AccountAPI.Service
@@ -51,17 +50,6 @@ namespace AccountAPI.Service
             var response = await _http.PutAsJsonAsync($"/api/Accounts/update-fullname/{id}", fullName);
             return response.IsSuccessStatusCode;
         }
-
-        public async Task<ChangePasswordResponse?> ChangePasswordAsync(ChangePasswordRequest request)
-        {
-            var response = await _http.PutAsJsonAsync("/api/Accounts/change-password", request);
-
-            if (!response.IsSuccessStatusCode)
-                return null;
-
-            return await response.Content.ReadFromJsonAsync<ChangePasswordResponse>();
-        }
-
 
 
     }
