+++ SEP490_Ezstay/AccountAPI/Service/Interfaces/IAuthApiClient.cs
@@ -1,5 +1,4 @@
 ﻿using AccountAPI.DTO.Response;
-using AccountAPI.DTO.Resquest;
 
 namespace AccountAPI.Service.Interfaces
 {
@@ -9,6 +8,5 @@ namespace AccountAPI.Service.Interfaces
         Task<bool> UpdateEmailAsync(string oldEmail, string newEmail);
         Task<AccountResponse?> GetByIdAsync(Guid id);
         Task<bool> UpdateFullNameAsync(Guid id, string fullName);
-        Task<ChangePasswordResponse?> ChangePasswordAsync(ChangePasswordRequest request);
      }
 }
