+++ SEP490_Ezstay/AuthApi/Controllers/AccountsController.cs
@@ -93,13 +93,5 @@ namespace AuthApi.Controllers
                 : NotFound("Không tìm thấy tài khoản");
         }
 
-        [HttpPut("change-password")]
-        public async Task<IActionResult> ChangePassword([FromBody] ChangePasswordRequest request)
-        {
-            var response = await _service.ChangePasswordAsync(request);
-            return response.Success ? Ok(response) : BadRequest(response);
-        }
-
-
     }
 }
