--- SEP490_Ezstay/AuthApi/DTO/Request/ChangePasswordRequest.cs
@@ -1,9 +1,0 @@
-﻿namespace AuthApi.DTO.Request
-{
-    public class ChangePasswordRequest
-    {
-        public string Email { get; set; } = null!;
-        public string OldPassword { get; set; } = null!;
-        public string NewPassword { get; set; } = null!;
-    }
-}
