+++ SEP490_Ezstay/AuthApi/DTO/Request/ResetPasswordRequestDto.cs
@@ -2,12 +2,7 @@
 {
     public class ResetPasswordRequestDto
     {
-     
-            public string Email { get; set; }        // Bổ sung email
-            public string Token { get; set; }        // OTP từ email
-            public string NewPassword { get; set; }  // Mật khẩu mới
-        
-
-     
+        public string Token { get; set; }
+        public string NewPassword { get; set; }
     }
 }
