--- SEP490_Ezstay/AuthApi/DTO/Response/ChangePasswordResponse.cs
@@ -1,9 +1,0 @@
-﻿namespace AuthApi.DTO.Response
-{
-    public class ChangePasswordResponse
-    {
-        public bool Success { get; set; }
-        public string Message { get; set; } = string.Empty;
-
-    }
-}
