+++ SEP490_Ezstay/AuthApi/Repositories/Interfaces/IAccountRepository.cs
@@ -12,5 +12,6 @@ namespace AuthApi.Repositories.Interfaces
         Task<Account?> UpdateAsync(Account account);
         Task MarkAsVerified(string email);
         Task BanAccountAsync(Guid id, bool isBanned);
+
     }
 }
