+++ SEP490_Ezstay/AuthApi/Services/AccountService.cs
@@ -73,26 +73,14 @@ namespace AuthApi.Services
 
             var currentRole = GetCurrentUserRole();
 
-            // Staff không được update tài khoản Admin
+            // Staff không update Admin
             if (currentRole == RoleEnum.Staff && acc.Role == RoleEnum.Admin)
                 return null;
-
-            // Chỉ Admin mới được đổi mật khẩu của tài khoản Staff
-            if (acc.Role == RoleEnum.Staff && currentRole != RoleEnum.Admin)
-            {
-                // Không thay đổi mật khẩu nếu không phải Admin
-                request.Password = null;
-            }
 
             acc.FullName = request.FullName;
             acc.Email = request.Email;
             acc.Phone = request.Phone;
-
-            // Hash lại mật khẩu nếu có
-            if (!string.IsNullOrWhiteSpace(request.Password))
-            {
-                acc.Password = BCrypt.Net.BCrypt.HashPassword(request.Password);
-            }
+            acc.Password = BCrypt.Net.BCrypt.HashPassword(request.Password);
 
             // Rule Role khi update
             acc.Role = request.Role switch
@@ -105,7 +93,6 @@ namespace AuthApi.Services
             var updated = await _repo.UpdateAsync(acc);
             return updated == null ? null : _mapper.Map<AccountResponse>(updated);
         }
-
 
         public async Task<AccountResponse> CreateAsync(AccountRequest request)
         {
@@ -126,67 +113,22 @@ namespace AuthApi.Services
             await _repo.CreateAsync(account);
             return _mapper.Map<AccountResponse>(account);
         }
-                public async Task<bool> UpdateFullNameAsync(Guid id, string fullName)
-                {           
-                     var acc = await _repo.GetByIdAsync(id);
-                if (acc == null) return false;
+public async Task<bool> UpdateFullNameAsync(Guid id, string fullName)
+{
+    var acc = await _repo.GetByIdAsync(id);
+    if (acc == null) return false;
 
-                acc.FullName = fullName; // ✅ chỉ cập nhật 1 field
+    acc.FullName = fullName; // ✅ chỉ cập nhật 1 field
 
-                     var updated = await _repo.UpdateAsync(acc);
-                 return updated != null;
-            }
+    var updated = await _repo.UpdateAsync(acc);
+    return updated != null;
+}
 
 
 
         public async Task VerifyAsync(string email) => await _repo.MarkAsVerified(email);
         public async Task BanAsync(Guid id) => await _repo.BanAccountAsync(id, true);
         public async Task UnbanAsync(Guid id) => await _repo.BanAccountAsync(id, false);
-
-
-        public async Task<ChangePasswordResponse> ChangePasswordAsync(ChangePasswordRequest request)
-        {
-            var acc = await _repo.GetByEmailAsync(request.Email); // <-- dùng Email
-            if (acc == null)
-            {
-                return new ChangePasswordResponse
-                {
-                    Success = false,
-                    Message = "Tài khoản không tồn tại"
-                };
-            }
-
-            // So sánh mật khẩu cũ
-            var isMatch = BCrypt.Net.BCrypt.Verify(request.OldPassword, acc.Password);
-            if (!isMatch)
-            {
-                return new ChangePasswordResponse
-                {
-                    Success = false,
-                    Message = "Mật khẩu cũ không đúng"
-                };
-            }
-
-            // Hash và lưu mật khẩu mới
-            acc.Password = BCrypt.Net.BCrypt.HashPassword(request.NewPassword);
-            var updated = await _repo.UpdateAsync(acc);
-
-            if (updated == null)
-            {
-                return new ChangePasswordResponse
-                {
-                    Success = false,
-                    Message = "Đổi mật khẩu thất bại"
-                };
-            }
-
-            return new ChangePasswordResponse
-            {
-                Success = true,
-                Message = "Đổi mật khẩu thành công"
-            };
-        }
-
     }
 
 }
