--- SEP490_Ezstay/BoardingHouseAPI/Services/Interfaces/IAuthService.cs
@@ -1,17 +1,0 @@
-﻿using AuthApi.DTO.Request;
-using AuthApi.DTO.Response;
-
-namespace AuthApi.Services.Interfaces
-{
-    public interface IAuthService
-    {
-        Task<RegisterResponseDto> RegisterAsync(RegisterRequestDto dto);
-        Task<LoginResponseDto> LoginAsync(LoginRequestDto dto);
-        Task<RegisterResponseDto> ResendOtpAsync(string email);
-        Task<RegisterResponseDto> CreateStaffAsync(CreateStaffRequestDto dto);
-        Task<RegisterResponseDto> ResetPasswordAsync(ResetPasswordRequestDto dto);
-        Task<RegisterResponseDto> ForgotPasswordAsync(ForgotPasswordRequestDto dto);
-        Task<RegisterResponseDto> SendPhoneOtpAsync(string phone);
-        Task<RegisterResponseDto> VerifyPhoneOtpAsync(string phone, string otp);
-        }
-}
