--- SEP490_Ezstay/BoardingHouseAPI/Services/Interfaces/IFacebookAuthService.cs
@@ -1,9 +1,0 @@
-﻿using AuthApi.Models;
-
-namespace AuthApi.Services.Interfaces
-{
-    public interface IFacebookAuthService
-    {
-        Task<Account> FacebookLoginAsync(string accessToken);
-    }
-}
