--- SEP490_Ezstay/BoardingHouseAPI/Services/Interfaces/IGoogleAuthService.cs
@@ -1,9 +1,0 @@
-﻿using AuthApi.Models;
-
-namespace AuthApi.Services.Interfaces
-{
-    public interface IGoogleAuthService
-    {
-        Task<Account> GoogleLoginAsync(string idToken);
-    }
-}
