--- SEP490_Ezstay/BoardingHouseAPI/Services/Interfaces/IPhoneVerificationService.cs
@@ -1,8 +1,0 @@
-﻿namespace AuthApi.Services.Interfaces
-{
-    public interface IPhoneVerificationService
-    {
-        Task SendOtpAsync(string phone);
-        Task<bool> VerifyOtpAsync(string phone, string otp);
-    }
-}
